@@ -1,13 +1,13 @@
 module github.com/ozontech/cute
 
-go 1.19
+go 1.21
 
 require (
-	github.com/josephburnett/jd v1.5.1
-	github.com/ohler55/ojg v1.19.4
-	github.com/ozontech/allure-go/pkg/allure v0.6.11
-	github.com/ozontech/allure-go/pkg/framework v0.6.28
-	github.com/stretchr/testify v1.7.1
+	github.com/josephburnett/jd v1.7.1
+	github.com/ohler55/ojg v1.21.1
+	github.com/ozontech/allure-go/pkg/allure v0.6.12
+	github.com/ozontech/allure-go/pkg/framework v0.6.29
+	github.com/stretchr/testify v1.8.4
 	github.com/xeipuuv/gojsonschema v1.2.0
 	moul.io/http2curl/v2 v2.3.0
 )
@@ -15,13 +15,14 @@ require (
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-openapi/jsonpointer v0.19.5 // indirect
-	github.com/go-openapi/swag v0.19.5 // indirect
+	github.com/go-openapi/swag v0.21.1 // indirect
 	github.com/google/uuid v1.3.0 // indirect
-	github.com/mailru/easyjson v0.0.0-20190626092158-b2ccc519800e // indirect
+	github.com/josharian/intern v1.0.0 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/xeipuuv/gojsonpointer v0.0.0-20180127040702-4e3ac2762d5f // indirect
 	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
